pipex_no_error

+++ libft/ft_free_tab.c
@@ -6,7 +6,7 @@
 /*   By: rbalazs <rbalazs@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/16 14:22:48 by rbalazs           #+#    #+#             */
-/*   Updated: 2024/08/16 14:23:03 by rbalazs          ###   ########.fr       */
+/*   Updated: 2024/08/23 18:55:30 by rbalazs          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_free_tab(char **tab)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (tab[i])
